requirements.txt
pandas>=2.0.0
streamlit>=1.30.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
scikit-learn>=1.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
